♻️ Document uids properly, document other submodules of base
